treewide: Align networking code with net_mgmt API change
The net_mgmt event representation changed from 32-bit to 64-bit value,
therefore update all net_mgmt callbacks to align with this change.

+++ lib/location/scan_wifi.c
@@ -172,7 +172,7 @@ static void scan_wifi_done_handle(struct net_mgmt_event_callback *cb)
 
 void scan_wifi_net_mgmt_event_handler(
 	struct net_mgmt_event_callback *cb,
-	uint32_t mgmt_event,
+	uint64_t mgmt_event,
 	struct net_if *iface)
 {
 	ARG_UNUSED(iface);
